refactor(cli): Rename cli related types module
